Adding Markdown support to contact forms
Uses the Mardown-it library to convert markdown formatted text from the user into HTML for submission as form data, currently to MS Flow.

This was necessary to preserve line break formatting in messages when data is moved between actions within the MS Flow (due to difficulties with literal line breaks in JSON data). HTML encodes line breaks etc. as proper elements which means the string itself can be a single line.

Adding markdown support is also useful for users that wish to include lists or other formatting.