Switching to external Python Version Parser project to generate the relevant version for the Python package and Sentry release. Coupled with the new inherited variables support, this means we can now calculate the version once and use this in any dependent jobs as needed.
Adding review and release container image tags so that images built in different circumstances can be managed
separately so that a tag retention policy can be applied. This ensures the number of images is kept manageable.

This means we don't need to build the deployment container on each commit and instead use per-release or per-review-app
tags that are rebuilt only as needed. We still have per-commit tags as well in case these are needed for testing, but
they will be limited in number. These images are now labelled as `build` tags meaning all reference to ‘deploy’ tags
should be removed.

Adding GitLab rules and needs to manage which tasks are ran in different situations (review apps vs. production). The
use of needs allows jobs to run without needing preceding stages to pass, if jobs don't need a job in an earlier stage.

Refactoring Terraform commands to make them more legible.