Adding `app-cli` Docker Compose service
Needed to allow the application container (running Flask debug server) to run in the background when an application command is being debugged in PyCharm.

Without this the CSW Server routes won’t work (as PyCharm would run the app service as a CLI only so nothing is listening for requests).